React Conversion - Teacher Roster
