+++ EdFi.Buzz.UI.Angular/src/app/Components/StudentCardReact/StudentCard.css
@@ -1,0 +1,21 @@
+.card-max-width{
+  max-width: 400px !important;
+}
+.bubble-counter {
+    top: 5px;
+    left: 5rem;
+}
+.img-size {
+  max-height: 70px;
+  max-width: 90px;
+}
+
+.image-container img {
+  width: 80%;
+  border-radius: 50%;
+  border: 5px solid lightgrey;
+}
+
+.card-footer {
+  padding:0;
+}
