+++ EdFi.Buzz.UI.Angular/src/app/Features/Landings/TeacherLandingReact/StudentTable.css
@@ -1,0 +1,4 @@
+.table-btn-outline{
+    display: inline-flex !important;
+    margin-right: 1rem;
+}
