+++ EdFi.Buzz.UI.Angular/src/app/Features/Landings/TeacherLandingReact/teacherLanding.css
@@ -6,16 +6,10 @@
     background-color: #007bff;
 }
 
-.imgage-round {
+.image-round {
     width: 70%;
     border-radius: 50%;
     border: 5px solid lightgrey;
-}
-
-
-.table-btn-outline{
-    display: inline-flex !important; 
-    margin-right: 1rem;
 }
 
 @media (max-width: 575.98px) {
